#4646: Dash, Route concepts added
