+++ dashboard/package.json
@@ -16,7 +16,8 @@
     "@mui/material": "^5.15.2",
     "axios": "^1.6.3",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.43",
