+++ dashboard/src/componets/Body/Body.jsx
@@ -1,27 +1,15 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
+import {BrowserRouter, Routes, Route} from "react-router-dom"
 import "./body.css"
-import axios from 'axios';
+import Dashboard from '../Dashboard/Dashboard'
+import Error from '../Error/Error'
 export default function Body() {
-  const [data,setData]=useState([])
-  useEffect(() => {
-    const fetchData = async () => {
-      const response = await axios.get('https://dummyjson.com/products');
-      setData(response.data.products);
-    };
-    fetchData();
-  }, []);
   return (
-    <div className='body_main'>
-        {
-          data.map((item)=>{
-            return(
-              <div className='card_head'>
-              <li>{item.title}</li>
-              <li>{item.description}</li>
-              </div>
-            )
-          })
-        }
-    </div>
+    <BrowserRouter>
+        <Routes>
+            <Route path = "/" element = {<Dashboard/>}/>
+            <Route path = 'error' element = {<Error/>}/>
+        </Routes>
+    </BrowserRouter>
   )
 }
