+++ dashboard/src/componets/Error/Error.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Error() {
+  return (
+    <div>
+        <h2>Error This Page Not Found</h2>
+    </div>
+  )
+}
